Refactor header components and enhance user authentication flow
- Updated `LandingHeader` component to accept user data for personalized navigation.
- Refactored `Header` component to include a link for the application title.
- Removed hover effect on anchor tags in global CSS for a cleaner design.
- Integrated Supabase user authentication in the home page for improved user experience.
