+++ app/components/landing-header.tsx
@@ -1,7 +1,11 @@
+"use client";
+
 import Link from "next/link";
 import { Button } from "@/components/shadcn/button";
+import { User } from "@supabase/supabase-js";
+import { logout } from "../(root)/action/logout";
 
-export default function LandingHeader() {
+export default function LandingHeader({ user }: { user?: User | null }) {
   return (
     <header className="fixed top-0 left-0 right-0 z-50 bg-background/80 backdrop-blur-md border-b border-border/40">
       <div className="container mx-auto px-4 sm:px-6 lg:px-8">
@@ -37,12 +41,25 @@ export default function LandingHeader() {
           </nav>
 
           <div className="flex items-center space-x-4">
-            <Button variant="ghost" asChild>
-              <Link href="/auth/login">Sign In</Link>
-            </Button>
-            <Button asChild>
-              <Link href="/auth/signup">Get Started</Link>
-            </Button>
+            {user ? (
+              <>
+                <Button variant="ghost" onClick={logout}>
+                  Sign Out
+                </Button>
+                <Button asChild>
+                  <Link href="/private">Get Started</Link>
+                </Button>
+              </>
+            ) : (
+              <>
+                <Button variant="ghost" asChild>
+                  <Link href="/auth/login">Sign In</Link>
+                </Button>
+                <Button asChild>
+                  <Link href="/auth/signup">Get Started</Link>
+                </Button>
+              </>
+            )}
           </div>
         </div>
       </div>
