+++ app/globals.css
@@ -66,10 +66,6 @@ a {
   cursor: pointer;
 }
 
-a:hover {
-  text-decoration: underline;
-}
-
 input {
   border: 1px solid #ccc;
   padding: 8px 16px;
